chore: delete empty v2/session-common + collapse patch barrel
- Delete packages/opencode/src/v2/session-common.ts (empty namespace,
  zero imports).
- Collapse patch/ barrel: move patch/patch.ts content into patch/index.ts,
  append self-reexport `export * as Patch from "."`, delete patch/patch.ts.

Both are cleanup as part of the namespace-treeshake migration.

Verified: tsgo --noEmit clean, SDK generate produces no drift,
relevant tests pass.
