--- packages/opencode/src/v2/session-common.ts
@@ -1,1 +1,0 @@
-export namespace SessionCommon {}
